Tests could be done (reference outputs needed)
test.sh should be good - it will still need to be edited once we add
more test targets (only scanner is supported for now)
We need test inputs and reference outputs for them
Basically, we have a tests folder, in which we have the test.sh script,
which we will execute when testing. There is also a makefile to compile
the desired tests (it will be called by test.h). Finally, there are
folders for each test target, eg. binary search tree, scanner and so
on.. In these folders, there should be a .c file that is used for
testing, inputs (in 'inputs' folder) and reference outputs (in
'reference' folder). After running the tests, outputs are saved in the
'outputs' folder.

Running the tests:
./test.sh target (for example 'scanner')
The output shows which tests passed and which did not - in case of an
unsuccessful tests, difference is generated (of output and reference)
and is also saved to logs/tests/target